Move pkgref to types
