+++ sandbox/runx/impl/types/pkgref_test.go
@@ -1,20 +1,20 @@
-package pkgref_test
+package types_test
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"go.jetpack.io/runx/impl/pkgref"
+	"go.jetpack.io/runx/impl/types"
 )
 
 func TestFromString_Valid(t *testing.T) {
 	testdata := []struct {
 		str string
-		ref pkgref.PkgRef
+		ref types.PkgRef
 	}{
 		{
 			str: "foo/bar",
-			ref: pkgref.PkgRef{
+			ref: types.PkgRef{
 				Owner:   "foo",
 				Repo:    "bar",
 				Version: "latest",
@@ -22,7 +22,7 @@ func TestFromString_Valid(t *testing.T) {
 		},
 		{
 			str: "foo/bar@v1.2.3",
-			ref: pkgref.PkgRef{
+			ref: types.PkgRef{
 				Owner:   "foo",
 				Repo:    "bar",
 				Version: "v1.2.3",
@@ -32,7 +32,7 @@ func TestFromString_Valid(t *testing.T) {
 
 	for _, td := range testdata {
 		t.Run(td.str, func(t *testing.T) {
-			ref, err := pkgref.FromString(td.str)
+			ref, err := types.NewPkgRef(td.str)
 			assert.NoError(t, err)
 			assert.Equal(t, td.ref, ref)
 		})
@@ -53,7 +53,7 @@ func TestFromString_Invalid(t *testing.T) {
 
 	for _, td := range testdata {
 		t.Run(td.str, func(t *testing.T) {
-			_, err := pkgref.FromString(td.str)
+			_, err := types.NewPkgRef(td.str)
 			assert.Error(t, err)
 		})
 	}
